refactor: Standalone file for FileIcon
