+++ src/viewer/NoViewer/NoViewer.jsx
@@ -1,22 +1,15 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { translate } from 'cozy-ui/react/I18n'
-import { Icon, Button, Alerter } from 'cozy-ui/transpiled/react'
+import { Button, Alerter } from 'cozy-ui/transpiled/react'
 import { logException } from 'drive/lib/reporter'
 import { isMobileApp } from 'cozy-device-helper'
 import CallToAction from './CallToAction'
+import FileIcon from './FileIcon'
 
 import { openLocalFileCopy } from 'drive/mobile/modules/offline/duck'
 
 import styles from '../styles.styl'
-import IconFiles from '../icons/icon-type-files.svg'
-import IconBin from '../icons/icon-type-bin.svg'
-import IconCode from '../icons/icon-type-code.svg'
-import IconSpreadsheet from '../icons/icon-type-sheet.svg'
-import IconSlide from '../icons/icon-type-slide.svg'
-import IconText from '../icons/icon-type-text.svg'
-import IconZip from '../icons/icon-type-zip.svg'
-import IconPdf from '../icons/icon-type-pdf.svg'
 
 class AsyncActionButton extends React.Component {
   state = {
@@ -82,39 +75,6 @@ const NoViewerButton = ({ file, fallbackUrl, t, onError }) => {
     return <OpenWithCordovaButton t={t} file={file} onError={onError} />
   else if (fallbackUrl) return <OpenWithWebButton t={t} url={fallbackUrl} />
   else return <DownloadButton t={t} file={file} />
-}
-
-const FileIcon = ({ type }) => {
-  let icon
-
-  switch (type) {
-    case 'bin':
-      icon = IconBin
-      break
-    case 'code':
-      icon = IconCode
-      break
-    case 'spreadsheet':
-      icon = IconSpreadsheet
-      break
-    case 'slide':
-      icon = IconSlide
-      break
-    case 'text':
-      icon = IconText
-      break
-    case 'zip':
-      icon = IconZip
-      break
-    case 'pdf':
-      icon = IconPdf
-      break
-    default:
-      icon = IconFiles
-      break
-  }
-
-  return <Icon icon={icon} width={160} height={140} />
 }
 
 class NoViewer extends React.Component {
